Refactor NNPInput imports to use dataset module
